+++ packages/editor/web/src/RichContentEditor/Toolbars/SideToolbar/AddPluginMenu.js
@@ -15,7 +15,8 @@ export default class AddPluginMenu extends Component {
     const { addPluginMenuConfig, isMobile, theme } = props;
     this.styles = mergeStyles({ styles: Styles, theme });
     this.showSearch = addPluginMenuConfig?.showSearch && !isMobile;
-    this.horizontalMenu = !addPluginMenuConfig && !isMobile;
+    this.horizontalMenu =
+      (!addPluginMenuConfig || addPluginMenuConfig?.horizontalMenuLayout) && !isMobile;
     this.wrapperClassName = classNames(this.styles.sideToolbarPanelWrapper, {
       [this.styles.horizontalMenu]: this.horizontalMenu,
     });
@@ -41,6 +42,7 @@ export default class AddPluginMenu extends Component {
       toolbarName,
       searchablePlugins,
     } = this.props;
+    const smallPlusIcon = addPluginMenuConfig?.tablePluginMenu;
     const { showSearch, wrapperClassName, pluginsClassName, horizontalMenu } = this;
     const { value } = this.state;
     return (
@@ -70,6 +72,7 @@ export default class AddPluginMenu extends Component {
             hidePopup={hidePopup}
             splitToSections={!value && addPluginMenuConfig?.splitToSections}
             horizontalMenu={horizontalMenu}
+            smallPlusIcon={smallPlusIcon}
             theme={theme}
             pluginMenuButtonRef={pluginMenuButtonRef}
             isMobile={isMobile}